MT-4 : Added ini4j library to make parsing the settings.ini.cfm file easier.
Tied in the data input through the new install settings page.  The data is now being passed to the settings.ini.cfm file! Woot!  My only concern about using the ini4j library is that it surrounds all the equal signs with spaces. I need to test and see if that is a problem for Mura.